Make default schenario name use the parent folder
Changes the default scenario name from being just hardcoded 'default'
to be the directory name hosting the molecule.yml file.

This change will allow us to reuse (symlink) molecule.yml file across
scenarios and roles because we will not longer be forced to define
this field in each of them.

Signed-off-by: Sorin Sbarnea <ssbarnea@redhat.com>
Related-Bug: #1744